requirements.txt
# Web Framework
streamlit==1.29.0

# HTTP Client (pinned for OpenAI compatibility)
httpx<0.28.0

# OpenAI API
openai==1.54.0

# PDF Processing
PyMuPDF==1.23.8
pdf2image==1.16.3

# OCR
pytesseract==0.3.10
Pillow==10.1.0

# Database
sqlalchemy==2.0.23

# Cloud Storage (Cloudflare R2)
boto3==1.35.0

# Environment Variables
python-dotenv==1.0.0
